feat(github): delete MasterFeaturesSection, add App-wide group, recompose tab account-first
